+++ packages/node/src/indexer/mmr.service.ts
@@ -5,16 +5,21 @@ import fs from 'fs';
 import { Injectable, OnApplicationShutdown } from '@nestjs/common';
 import { u8aToHex, u8aEq } from '@polkadot/util';
 import { DEFAULT_WORD_SIZE, DEFAULT_LEAF, MMR_AWAIT_TIME } from '@subql/common';
+import { NodeConfig } from '@subql/node-core/configure';
+import { MmrPayload, MmrProof } from '@subql/node-core/events';
+import {
+  MetadataFactory,
+  MetadataRepo,
+  PoiFactory,
+  PoiRepo,
+  ProofOfIndex,
+} from '@subql/node-core/indexer/entities';
+import { delay } from '@subql/node-core/utils';
 import { MMR, FileBasedDb } from '@subql/x-merkle-mountain-range';
 import { keccak256 } from 'js-sha3';
 import { Sequelize, Op } from 'sequelize';
-import { NodeConfig } from '@subql/node-core/configure';
 import { SubqueryProject } from '../configure/SubqueryProject';
 import { getLogger } from '../utils/logger';
-import { delay } from '@subql/node-core/utils';
-import { MetadataFactory, MetadataRepo } from '@subql/node-core/indexer/entities';
-import { PoiFactory, PoiRepo, ProofOfIndex } from '@subql/node-core/indexer/entities';
-import { MmrPayload, MmrProof } from './events';
 const logger = getLogger('mmr');
 
 const DEFAULT_FETCH_RANGE = 100;
